Copy CompileFile1 into targets & unnest CompileFileLine
The latter is just because the extra later of nesting seems so
redundant.

The former is in prep of virtualising them.  I'd've preferred to keep
the original paths, but the entry point of Dotty we're using takes
"fileNames: List[String]", and it's a chunky piece of code to copy +
adapt to our own SourceFile/VirtualFile abstraction.